훅 및 컴포 분리
