--- src/pages/home/Home.tsx
@@ -1,12 +1,0 @@
-import { useState } from 'react';
-
-import DashboardMain from './DashboardMain';
-import LandingMain from './LandingMain';
-
-const Home = () => {
-  const [isLoggedIn] = useState(false);
-
-  return <>{isLoggedIn ? <DashboardMain /> : <LandingMain />}</>;
-};
-
-export default Home;
